pyproject.toml
[tool.poetry]
name = "yt-downloader-bot"
version = "0.1.0"
description = ""
authors = ["Alexander Gorshkov <agorshkov@mznx.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
aiogram = "^2.24"
pytube = "^12.1.2"
celery = {extras = ["redis"], version = "^5.2.7"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
